Fixed: init

+++ src/da_dev/Dev.cr
@@ -7,11 +7,43 @@ module DA_Dev
     def init
       repo_name = File.basename(Dir.current)
       shard_name = File.basename(Dir.current, ".cr")
+      init_bin(shard_name, repo_name)
       init_gitignore
       Dir.mkdir_p("src")
       Dir.mkdir_p("specs")
       init_shard_yml(shard_name, repo_name)
     end
+
+    def init_bin(shard_name, repo_name)
+      Dir.mkdir_p("bin")
+      file = "bin/__.cr"
+      default_contents = <<-EOF
+
+      THIS_DIR = File.dirname(__DIR__)
+      require "da_dev"
+      full_cmd = ARGV.join(" ")
+      args     = ARGV.dup
+      cmd      = args.shift
+
+      case
+
+      when "-h --help help".split.includes?(full_cmd)
+        # === {{CMD}} -h|--help|help
+        DA_Dev::Documentation.print_help([__FILE__])
+
+      else
+        red! "!!! Invalid arguments: \#{ARGV.map(&.inspect).join " "}"
+        exit 1
+
+      end # === case
+
+      EOF
+
+      if !File.exists?(file)
+        File.write(file, default_contents)
+        DA_Dev.green! "=== BOLD{{Wrote}}: {{#{file}}}"
+      end
+    end # === def init_bin
 
     def init_shard_yml(shard_name, repo_name)
       default_contents = <<-EOF
@@ -36,20 +68,24 @@ module DA_Dev
 
     def init_gitignore
       file = ".gitignore"
+      old_contents = ""
       contents = if File.exists?(file)
-                   File.read(file).strip.split("\n")
+                   old_contents = File.read(file).strip
+                   old_contents.split("\n")
                  else
                    [] of String
                  end
       contents = contents.concat(%w[/tmp/ /.js_packages/ /shard.lock /.shards/]).sort.uniq
       contents.push("")
       contents = contents.join('\n')
-      is_new = File.exists?(file)
-      File.write(file, contents)
-      if is_new
+      if !File.exists?(file)
+        File.write(file, contents)
         DA_Dev.green! "=== BOLD{{Wrote}}: {{#{file}}}"
       else
-        DA_Dev.green! "=== BOLD{{Updated}}: {{#{file}}}"
+        if old_contents.strip != contents.strip
+          File.write(file, contents)
+          DA_Dev.green! "=== BOLD{{Updated}}: {{#{file}}}"
+        end
       end
     end
